Use path hashes of parameters instead of custom build property
The `path_hashes` hve been extracted from the
`CustomBuildPropertiesAction` of a job. This value was no longer
updated and thereby causing mismatching builds for all builds since the
last update of the build property.
With this fix the `path_hashes` are extraced the same way as all other
build parameters. This now finds matching builds again

CMK-24367

Change-Id: I4de7417c5745979280989426f09f98f85d28fef2