Return empty artifacts list instead of raising an error
Previously an error "Job has no artifacts!" was raised if the download
did not return any artifacts in case the job had no artifacts.
With this change an empty list is returned. The decision to raise an
error if nothing is provided by the job should be up to the consumer,
but not to the tool.

CMK-26116

Change-Id: I0c3a23e209ad479dd959de0b735532743f0494e8