Do not raise an error when awaiting a build or downloading an artifact of a failed build
Before this change an error was raised if the job in question did not
pass successfully. Downloading artifacts of that job was not possible.

With this change the hard error is removed and the response JSON is
changed in case the awaited build or the job to download artifacts has
the job status as anything but success.
This is now in sync with the result returned when requesting a
successfully passed job.

CMK-26128

Change-Id: I00f2c671ad9229de17baa9ab4fed583bb0bf8ff7