Do not overwrite build result with older running result
While iterating over all builds for a result of an InfluxDB query, the
build result might have been overwritten by an older (in time) result.
This causes the response to become
`{"existing": {"result": "RUNNING", ...}}`.
The consumer task would consider this job as actively running and
continue with a await call. As the job already finished the awaited
task would return `FAILURE` as this is the final, already known, result.
Finally this `FAILURE` result would not cause a new build, but failing
instantly, being caught in a somehow endless loop, without the
possibility to escape other than with a new commit.

CMK-26845

Change-Id: I42a9f82819101253f249b9a948dd6619291d9fb2