Do not overwrite build parameters by an older InfluxDB entry
During high load situations on Jenkins the job parameter of a build can
not be extracted properly by the Groovy code. The iterator will fall
back to default values like `null` or `false` and post these to the
InfluxDB.

When iterating over the entries of an InfluxDB query the latest entry
is processed first, see CMK-26742 and release `1.0.1`. The `Build`
object is populated with the correct Jenkins job parameters of the
`SUCCESS` entry and overwritten to `null` or `false` on the second
iteration while processing the `RUNNING` entry.

This change fixes this specific Jenkins load issue by not overwriting
the parameters of `Build` when processing the `RUNNING` InfluxDB entry.

CMK-28117

Change-Id: I032a0dcc91806eb2b274766e4dd1c0158181a972